+++ online-exam-platform/src/main/java/com/iuc/cerrahpasa/onlineexamplatform/service/EmailService.java
@@ -1,28 +1,9 @@
 package com.iuc.cerrahpasa.onlineexamplatform.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.mail.SimpleMailMessage;
-import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.stereotype.Service;
 
 @Service
-public class EmailService {
+public interface EmailService {
 
-	@Autowired
-	private JavaMailSender emailSender;
-	
-	public void sendSimpleMessage(String to, String subject, String text) {
-        SimpleMailMessage message = new SimpleMailMessage();
-
-        try {
-        	message.setFrom("yilmazlpren@gmail.com");
-        	message.setTo(to);
-            message.setSubject(subject);
-            message.setText(text);
-            emailSender.send(message);
-        }catch(Exception e) {
-        	System.out.println(e.getLocalizedMessage());
-        }
-        
-    }
+	public void sendSimpleMessage(String to, String subject, String text);
 }
