+++ online-exam-platform/src/main/java/com/iuc/cerrahpasa/onlineexamplatform/service/impl/EmailServiceImpl.java
@@ -1,0 +1,35 @@
+package com.iuc.cerrahpasa.onlineexamplatform.service.impl;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.mail.SimpleMailMessage;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.stereotype.Service;
+
+import com.iuc.cerrahpasa.onlineexamplatform.service.EmailService;
+
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+@Service
+public class EmailServiceImpl implements EmailService {
+
+	@Autowired
+	private JavaMailSender emailSender;
+
+	@Override
+	public void sendSimpleMessage(String to, String subject, String text) {
+		SimpleMailMessage message = new SimpleMailMessage();
+
+		try {
+			message.setFrom("yilmazlpren@gmail.com");
+			message.setTo(to);
+			message.setSubject(subject);
+			message.setText(text);
+			emailSender.send(message);
+		} catch (Exception e) {
+			System.out.println(e.getLocalizedMessage());
+		}
+
+	}
+
+}
